feat: add a status field to group_memberships
A group admin should be able to edit the deployment and invite others to
join the group.
